+++ lib/utility/arguments.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class CardArguments {
+  Map<String, dynamic> cards;
+  Color colorOfCard;
+
+  CardArguments(this.cards, this.colorOfCard);
+}