+++ CHANGELOG.md
@@ -20,10 +20,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add login Skip
 - Add page step one for sharing a list
 - Add page step two for sharing a list
+- Add page for a big loyalty card
 
 ### Changed
 
 - Update share page
+- Update loyalty cards page
 
 ### Deprecated